+++ common/src/constants.ts
@@ -59,8 +59,7 @@ export const DERIVATIVES_TRADING_PORTFOLIO_MARGIN_REST_API_TESTNET_URL =
     'https://testnet.binancefuture.com';
 
 // Derivatives Trading (Portfolio Margin Pro) constants
-export const DERIVATIVES_TRADING_PORTFOLIO_MARGIN_PRO_REST_API_PROD_URL =
-    'https://api.binance.com';
+export const DERIVATIVES_TRADING_PORTFOLIO_MARGIN_PRO_REST_API_PROD_URL = 'https://api.binance.com';
 
 // Dual Investment constants
 export const DUAL_INVESTMENT_REST_API_PROD_URL = 'https://api.binance.com';